requirements.txt
tidalapi
python-dotenv